Changed timeElapsed to int as we're measuring time in seconds. Updated StringFormatter methods to reflect this change. Removed Model interface from RunContract as it is not needed

+++ app/src/main/java/com/tjrushby/runlite/contracts/DetailsContract.java
@@ -106,7 +106,7 @@ public interface DetailsContract {
 
         void onEditTextTimeElapsedClicked();
 
-        void onEditTextTimeElapsedUpdated(long timeElapsed);
+        void onEditTextTimeElapsedUpdated(int timeElapsed);
 
         void onImageViewFullscreenClicked();
