+++ app/src/main/java/com/tjrushby/runlite/dialogs/TimePickerDialogListener.java
@@ -1,5 +1,5 @@
 package com.tjrushby.runlite.dialogs;
 
 public interface TimePickerDialogListener {
-    void onTimePickerDialogPositiveClick(long timeElapsed);
+    void onTimePickerDialogPositiveClick(int timeElapsed);
 }
